changes to create pages of stocks and investments

+++ resources/views/investments/index.blade.php
@@ -19,6 +19,19 @@
         </tr>
         </thead>
         <tbody>
+		
+		<script>
+
+			function ConfirmDelete()
+				{
+					var x = confirm("Are you sure you want to delete? Click confirm to continue");
+					if (x)
+					return true;
+					else
+					return false;
+				}
+		</script>
+		
         @foreach ($investments as $investment)
             <tr>
                 <td>{{ $investment->customer->cust_number }}</td>
@@ -32,7 +45,7 @@
                 <td><a href="{{url('investments',$investment->id)}}" class="btn btn-primary">Read</a></td>
                 <td><a href="{{route('investments.edit',$investment->id)}}" class="btn btn-warning">Update</a></td>
                 <td>
-                    {!! Form::open(['method' => 'DELETE', 'route'=>['investments.destroy', $investment->id]]) !!}
+                    {!! Form::open(['method' => 'DELETE', 'route'=>['investments.destroy', $investment->id], 'onsubmit' => 'return ConfirmDelete()']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                     {!! Form::close() !!}
                 </td>
